@@ -1,8 +1,12 @@
 package org.example;
 
-import static org.junit.Assert.assertTrue;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Unit test for simple App.
@@ -13,8 +17,14 @@ public class AppTest
      * Rigorous Test :-)
      */
     @Test
-    public void shouldAnswerWithTrue()
+    @DisplayName("Single test successful")
+    public void shouldAnswerWithTrue() {
+        Assertions.assertTrue( 2 < 3 );
+    }
+
+    @Test
+    public void shouldAnswerWithFalse()
     {
-        assertTrue( true );
+        Assertions.assertFalse( 2 > 3 );
     }
 }
